moved Label to deprecated folder, and created fake Label pointer for backward compatibility
