+++ packages/wix-ui-core/src/components/deprecated/Label/Label.e2e.ts
@@ -1,7 +1,7 @@
 import * as eyes from 'eyes.it';
 import {browser} from 'protractor';
 import {getStoryUrl, waitForVisibilityOf} from 'wix-ui-test-utils/protractor';
-import {labelTestkitFactory} from '../../testkit/protractor';
+import {labelTestkitFactory} from '../../../testkit/protractor';
 
 describe('Label', () => {
   const storyUrl = getStoryUrl('Components', 'Label');
