+++ packages/wix-ui-core/src/components/deprecated/Label/Label.spec.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import {labelDriverFactory} from './Label.driver';
-import {ReactDOMTestContainer} from '../../../test/dom-test-container';
+import {ReactDOMTestContainer} from '../../../../test/dom-test-container';
 import {isEnzymeTestkitExists} from 'wix-ui-test-utils/enzyme';
 import {isTestkitExists} from 'wix-ui-test-utils/vanilla';
-import {labelTestkitFactory} from '../../testkit';
-import {labelTestkitFactory as enzymeLabelTestkitFactory} from '../../testkit/enzyme';
+import {labelTestkitFactory} from '../../../testkit';
+import {labelTestkitFactory as enzymeLabelTestkitFactory} from '../../../testkit/enzyme';
 import {Label} from './Label';
 import {mount} from 'enzyme';
