+++ packages/wix-ui-core/src/components/deprecated/Label/index.ts
@@ -1,0 +1,1 @@
+export {Label, LabelProps} from './Label';
